[MIG] crm_facebook_leads: migrate to 15.0
The following is performed:
- Rename field `odoo_field` on model `crm.facebook.form.field` to add
  missing `_id` suffix, as it's a many2one field
- Remove "Facebook Add" and "Facebook Campaign" from the lead model, as
  those fields always hold the same values as lead's medium and
  campaign, respectively
- Remove `.pot` file, as it's not actually being used because we handle
  translations manually
