+++ crm_facebook_leads/i18n/fr.po
@@ -4,11 +4,11 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 14.0+e\n"
+"Project-Id-Version: Odoo Server 15.0+e\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-12-10 00:42+0000\n"
-"PO-Revision-Date: 2020-12-10 00:42+0000\n"
-"Last-Translator: \n"
+"POT-Creation-Date: 2022-06-07 02:51+0000\n"
+"PO-Revision-Date: 2022-06-07 02:51+0000\n"
+"Last-Translator: Luis González <lgonzalez@vauxoo.com>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -16,9 +16,21 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: crm_facebook_leads
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_adset__id_facebook_adset
+msgid "Adset ID"
+msgstr ""
+
+#. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form__allow_to_sync
 msgid "Allow To Sync"
 msgstr "Autoriser La Synchronisation"
+
+#. module: crm_facebook_leads
+#: model:ir.actions.server,name:crm_facebook_leads.crm_facebook_cron_ir_actions_server
+#: model:ir.cron,cron_name:crm_facebook_leads.crm_facebook_cron
+#: model:ir.cron,name:crm_facebook_leads.crm_facebook_cron
+msgid "CRM Facebook Lead Ads: Fetch leads"
+msgstr "CRM Facebook Lead Ads: Obtenir des Prospects Facebook"
 
 #. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form__campaign_id
@@ -50,24 +62,19 @@ msgid "Display Name"
 msgstr "Nom Affiché"
 
 #. module: crm_facebook_leads
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_ad_id
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_social_account__id_facebook_ad
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_medium__id_facebook_ad
-msgid "Facebook Ad"
-msgstr "Annonce Facebook"
+msgid "Facebook Ad ID"
+msgstr "ID D'annonce Facebook"
 
 #. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_adset_id
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_adset__facebook_adset_id
 msgid "Facebook Adset"
 msgstr "Groupe D'Annonces Facebook"
 
 #. module: crm_facebook_leads
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_campaign_id
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_marketing_campaign__facebook_campaign_id
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_campaign__id_facebook_campaign
-msgid "Facebook Campaign"
-msgstr "Campagne Facebook"
+msgid "Facebook Campaign ID"
+msgstr "ID de campagne Facebook"
 
 #. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_date_create
@@ -83,6 +90,11 @@ msgstr "Champ Facebook"
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_form_id
 msgid "Facebook Form"
 msgstr "Formulaire Facebook"
+
+#. module: crm_facebook_leads
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form__id_facebook_form
+msgid "Facebook Form ID"
+msgstr "ID de formulaire Facebook"
 
 #. module: crm_facebook_leads
 #: model:ir.model,name:crm_facebook_leads.model_crm_facebook_form
@@ -106,16 +118,15 @@ msgid "Facebook Is Organic"
 msgstr "Est Organique Facebook"
 
 #. module: crm_facebook_leads
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_lead_id
-msgid "Facebook Lead"
-msgstr "Prospect Facebook"
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__id_facebook_lead
+msgid "Facebook Lead ID"
+msgstr "ID de prospect Facebook"
 
 #. module: crm_facebook_leads
 #: model:ir.model,name:crm_facebook_leads.model_crm_facebook_page
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form__page_id
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_page_id
 #: model_terms:ir.ui.view,arch_db:crm_facebook_leads.crm_facebook_page_form
-#: model_terms:ir.ui.view,arch_db:crm_facebook_leads.crm_facebook_page_tree
 msgid "Facebook Page"
 msgstr "Page Facebook"
 
@@ -136,21 +147,9 @@ msgid "Facebook form fields"
 msgstr "Champs de Formulaire Facebook"
 
 #. module: crm_facebook_leads
-#: model:ir.actions.server,name:crm_facebook_leads.crm_facebook_cron_ir_actions_server
-#: model:ir.cron,cron_name:crm_facebook_leads.crm_facebook_cron
-#: model:ir.cron,name:crm_facebook_leads.crm_facebook_cron
-msgid "Fetch Facebook Leads"
-msgstr "Obtenir des Prospects Facebook"
-
-#. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form_field__form_id
 msgid "Form"
 msgstr "Formulaire"
-
-#. module: crm_facebook_leads
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form__id_facebook_form
-msgid "Form ID"
-msgstr "ID Formulaire"
 
 #. module: crm_facebook_leads
 #: model_terms:ir.ui.view,arch_db:crm_facebook_leads.crm_facebook_form_form
@@ -168,7 +167,7 @@ msgstr "Obtenir des Formulaires"
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_page__id
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_adset__id
 msgid "ID"
-msgstr "ID"
+msgstr ""
 
 #. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form____last_update
@@ -229,7 +228,7 @@ msgid "Name"
 msgstr "Nom"
 
 #. module: crm_facebook_leads
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form_field__odoo_field
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form_field__odoo_field_id
 msgid "Odoo Field"
 msgstr "Champ Odoo"
 
@@ -275,13 +274,14 @@ msgid ""
 "This is a name that helps you keep track of your different campaign efforts,"
 " e.g. Fall_Drive, Christmas_Special"
 msgstr ""
-"C'est un nom qui vous aide à suivre vos différents efforts de campagne,"
-" par exemple Fall_Drive, Christmas_Special"
+"C'est un nom qui vous aide à suivre vos différents efforts de campagne, par "
+"exemple Fall_Drive, Christmas_Special"
 
 #. module: crm_facebook_leads
 #: model:ir.model.fields,help:crm_facebook_leads.field_crm_lead__facebook_ad_id
 msgid "This is the method of delivery, e.g. Postcard, Email, or Banner Ad"
-msgstr "C'est la méthode de livraison, par exemple Postcard, Email, or Banner Ad"
+msgstr ""
+"C'est la méthode de livraison, par exemple Postcard, Email, or Banner Ad"
 
 #. module: crm_facebook_leads
 #: model_terms:ir.ui.view,arch_db:crm_facebook_leads.crm_facebook_form_form
