+++ crm_facebook_leads/models/__init__.py
@@ -1,4 +1,5 @@
 from . import crm_facebook_form
+from . import crm_facebook_form_field
 from . import crm_facebook_page
 from . import crm_lead
 from . import utm_adset
