+++ crm_facebook_leads/models/crm_facebook_form.py
@@ -18,7 +18,7 @@ class CrmFacebookForm(models.Model):
 
     name = fields.Char(required=True)
     allow_to_sync = fields.Boolean()
-    id_facebook_form = fields.Char(string='Form ID', required=True, readonly=True)
+    id_facebook_form = fields.Char(string='Facebook Form ID', required=True, readonly=True)
     access_token = fields.Char(required=True, related='page_id.access_token', string='Page Access Token')
     page_id = fields.Many2one('crm.facebook.page', readonly=True, ondelete='cascade', string='Facebook Page')
     mappings = fields.One2many('crm.facebook.form.field', 'form_id')
@@ -40,20 +40,3 @@ class CrmFacebookForm(models.Model):
                 'name': qualifier['label'],
                 'facebook_field': qualifier.get('key', False) or qualifier.get('field_key', False)
             })
-
-
-class CrmFacebookFormField(models.Model):
-    _name = 'crm.facebook.form.field'
-    _description = 'Facebook form fields'
-
-    form_id = fields.Many2one('crm.facebook.form', required=True, ondelete='cascade', string='Form')
-    name = fields.Char()
-    odoo_field = fields.Many2one(
-        'ir.model.fields',
-        domain=[('model', '=', 'crm.lead'), ('store', '=', True), ('ttype', 'in', (
-            'char', 'date', 'datetime', 'float', 'html', 'integer', 'monetary', 'many2one', 'selection', 'phone',
-            'text'))], required=False)
-    facebook_field = fields.Char(required=True)
-
-    _sql_constraints = [
-        ('field_unique', 'unique(form_id, odoo_field, facebook_field)', 'Mapping must be unique per form')]
