+++ crm_facebook_leads/views/crm_facebook_form_views.xml
@@ -29,7 +29,7 @@
                         <tree editable="bottom">
                             <field name="name"/>
                             <field name="facebook_field"/>
-                            <field name="odoo_field"/>
+                            <field name="odoo_field_id"/>
                         </tree>
                     </field>
                 </group>
